@@ -1,14 +1,23 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:snow_weather_info/data/sources/avalanche_api.dart';
-import 'package:snow_weather_info/data/sources/data_api.dart';
+import 'package:shared_preferences/shared_preferences.dart';
+import 'package:snow_weather_info/core/notifier/preference.dart';
+import 'package:snow_weather_info/core/theme/app_theme.dart';
 import 'package:snow_weather_info/data/data_notifier.dart';
+import 'package:snow_weather_info/data/repositories/stations.dart';
+import 'package:snow_weather_info/data/sources/avalanche_api.dart';
 import 'package:snow_weather_info/data/sources/database_helper.dart';
 import 'package:snow_weather_info/data/sources/preferences.dart';
+import 'package:snow_weather_info/data/sources/station_api.dart';
 import 'package:snow_weather_info/ui/home_page.dart';
 
-void main() {
-  runApp(const MyApp());
+Future<void> main() async {
+  WidgetsFlutterBinding.ensureInitialized();
+  final prefs = await SharedPreferences.getInstance();
+
+  runApp(MyApp(
+    preferences: prefs,
+  ));
 }
 
 const _title = 'Info Neige';
@@ -16,60 +25,55 @@ const _title = 'Info Neige';
 class MyApp extends StatelessWidget {
   const MyApp({
     Key? key,
+    required this.preferences,
   }) : super(key: key);
+
+  final SharedPreferences preferences;
 
   @override
   Widget build(BuildContext context) {
     return MultiProvider(
-        providers: [
-          Provider<DatabaseHelper>(create: (_) => DatabaseHelper()),
-          Provider<AvalancheAPI>(create: (_) => AvalancheAPI()),
-          Provider<Preferences>(create: (_) => Preferences()),
-          Provider<DataAPI>(create: (_) => DataAPI()),
-          ChangeNotifierProxyProvider0<DataNotifier>(
-            create: (_) => DataNotifier(),
-            update: (context, old) => old!
-              ..avalancheAPI = context.watch<AvalancheAPI>()
-              ..dataAPI = context.watch<DataAPI>()
-              ..preferences = context.watch<Preferences>()
-              ..databaseHelper = context.watch<DatabaseHelper>()
-              ..initData(),
+      providers: [
+        Provider<DatabaseHelper>(create: (_) => DatabaseHelper()),
+        Provider<AvalancheAPI>(create: (_) => AvalancheAPI()),
+        Provider<Preferences>(create: (_) => Preferences(preferences)),
+        Provider<StationAPI>(create: (_) => StationAPI()),
+        ProxyProvider0<StationsRepository>(
+          update: (context, old) => StationsRepository.update(
+            context.watch<StationAPI>(),
+            context.watch<DatabaseHelper>(),
+            context.watch<Preferences>(),
+            old,
           ),
-        ],
-        child: MaterialApp(
+        ),
+        ChangeNotifierProxyProvider0<PreferenceNotifier>(
+          create: (_) => PreferenceNotifier(),
+          update: (context, old) => old!
+            ..preferences = context.watch<Preferences>()
+            ..init(),
+        ),
+        ChangeNotifierProxyProvider0<DataNotifier>(
+          create: (_) => DataNotifier(),
+          update: (context, old) => old!
+            ..avalancheAPI = context.watch<AvalancheAPI>()
+            ..stationAPI = context.watch<StationAPI>()
+            ..preferences = context.watch<Preferences>()
+            ..databaseHelper = context.watch<DatabaseHelper>()
+            ..stationsRepository = context.watch<StationsRepository>()
+            ..initData(),
+        ),
+      ],
+      child: Consumer<PreferenceNotifier>(
+          builder: (context, PreferenceNotifier notifier, child) {
+        return MaterialApp(
           title: _title,
-          theme: ThemeData(
-            brightness: Brightness.light,
-            fontFamily: 'Helvetica',
-            primaryColor: Colors.blueAccent,
-            accentColor: Colors.blueAccent,
-            backgroundColor: Colors.white,
-            disabledColor: Colors.grey,
-            textTheme: const TextTheme(
-              headline6: TextStyle(color: Colors.white),
-              bodyText2: TextStyle(color: Colors.black),
-              bodyText1: TextStyle(color: Colors.blue),
-              subtitle1: TextStyle(color: Colors.black),
-              subtitle2: TextStyle(color: Colors.black),
-            ),
-          ),
-          darkTheme: ThemeData(
-            brightness: Brightness.dark,
-            fontFamily: 'Helvetica',
-            primaryColor: Colors.blueAccent.shade400,
-            accentColor: Colors.blueAccent.shade400,
-            backgroundColor: Colors.grey.shade800,
-            disabledColor: Colors.grey.shade900,
-            textTheme: TextTheme(
-              headline6: const TextStyle(color: Colors.white),
-              bodyText2: const TextStyle(color: Colors.white70),
-              bodyText1: TextStyle(color: Colors.blue.shade300),
-              subtitle1: const TextStyle(color: Colors.white),
-              subtitle2: const TextStyle(color: Colors.white),
-            ),
-          ),
+          theme: AppTheme.ligthTheme,
+          darkTheme: AppTheme.darkTheme,
+          themeMode: notifier.themeMode,
           home: const MyHomePage(title: _title),
-        ));
+        );
+      }),
+    );
   }
 }
 
